+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Si vous souhaitez voir le détail, cliquez sur les numéros qui vous renverrons 
 # [2.12] - 
 ## Améliorations :
 - [#393](https://github.com/Refhi/Weda-Helper/issues/393) - Alt+chiffre ouvre désormais le dernier document, et Alt+Maj+chiffre en crée un nouveau (ex. Pour les ordonnances, les prescriptions, les consultations, etc.). Les demandes de nouvelles prescriptions sont mieux prises en compte.
+- [#390](https://github.com/Refhi/Weda-Helper/issues/390) - La case "Mon patient accepte que je transmette le présent avis d'arrêt de travail pour son compte [...]" se coche automatiquement.
 
 # [2.11.3] - Amélioration de l'apparence & Tipeee
 ## Améliorations :